### 2019-08-18 e2ab720
fixes to build and for language
for people without orcids, just remove the orcid part of the metadata...
