+++ paper.md
@@ -11,13 +11,11 @@ authors:
     orcid: 0000-0003-0872-7098
     affiliation: 1
   - name: Stefan Uhlich
-    orcid: 0000-0000-0000-0000
     affiliation: 2
   - name: Antoine Liutkus
     orcid: 0000-0002-3458-6498
     affiliation: 1
   - name: Yuki Mitsufuji
-    orcid: 0000-0000-0000-0000
     affiliation: 3
 affiliations:
  - name: Inria and LIRMM, University of Montpellier, France
@@ -27,7 +25,6 @@ affiliations:
  - name: Sony Corporation, Japan
    index: 3
 date: 17 July 2019
-
 bibliography: paper.bib
 ---
 
@@ -39,56 +36,56 @@ Music separation has a long history of scientific activity as it is known to be 
 In recent years, deep learning-based systems - for the first time - yielded high-quality separations that also lead to increased commercial interest.
 However, until now, no open-source implementation that achieves state-of-the-art results is available.
 _Open-Unmix_ closes this gap by providing a reference implementation based on deep neural networks.
-It serves two main purposes: Firstly, to accelerate academic research as _Open-Unmix_ provides implementations for the most popular deep learning frameworks, giving researchers a flexible way to reproduce results; Secondly, we provide a pre-trained model for end users and even artists to try and use source separation.
+It serves two main purposes. Firstly, to accelerate academic research as _Open-Unmix_ provides implementations for the most popular deep learning frameworks, giving researchers a flexible way to reproduce results. Secondly, we provide a pre-trained model for end users and even artists to try and use source separation.
 Furthermore, we designed _Open-Unmix_ to be one core component in an open ecosystem on music separation, where we already provide open datasets, software utilities, and open evaluation to foster reproducible research as the basis of future development.
 
 # Background
 
-Music separation is a problem which fascinated researchers for over 50 years. This is partly because, mathematically, there exists no closed-form solution when many sources (instruments) are recorded a mono or stereo signal.
-To address the problem, researchers exploited further knowledge about the way the signals were recorded and mixed. A large number of these methods are centered around "classical" signal processing methods. For a more detailed overview see [@rafii17] and [@cano19].
+Music separation is a problem which has fascinated researchers for over 50 years. This is partly because, mathematically, there exists no closed-form solution when many sources (instruments) are recorded in a mono or stereo signal.
+To address the problem, researchers exploited additional knowledge about the way the signals were recorded and mixed. A large number of these methods are centered around "classical" signal processing methods. For a more detailed overview see [@rafii17] and [@cano19].
 <!-- - Explain how and when deep neural network-based music separation outperformed existing methods. -->
 Many of these methods were hand-crafted and tuned to a small number of music recordings [@sisec10; @sisec11; @sisec13].
 Systematic objective evaluation of these methods, however, was hardly feasible as freely available datasets did not exist at that time.
 In fact, for a meaningful evaluation, the ground truth separated stems are necessary.
 However, these are considered precious assets in the music recording industry, therefore, are usually not available because commercial music is well known to be subject to copyright protection laws.
 
-Nonetheless, in the past five years, freely available datasets were released that enabled the development of data-driven methods thanks to some artists that choose compatible licenses like Creative Commons.
-Since then, progress in performance was closely linked to the availability of more data that allowed the use of machine learning-based methods.
-This lead to a large performance boost similar to other audio tasks such as automatic speech recognition (ASR) where a large amount of data was available.
-In fact, in 2016 the speech recognition community had access to datasets with more than 10.000 hours of speech [@amodei16].
+Nonetheless, in the past five years, freely available datasets were released that enabled the development of data-driven methods thanks to some artists who choose compatible licenses like Creative Commons.
[…]
