### 2019-08-18 #1659 @danielskatz
Please merge https://github.com/sigsep/open-unmix-paper-joss/pull/1 and see if that enables the paper to build by entering ```@whedon generate pdf``` as a new comment here.  (also note that I suggested a number of language changes as well in the PR - these are all optional...
