### 2019-08-18 #1659 @faroit
@danielskatz regarding the use of submodules, please note my questions I asked earlier in [JOSS#565](https://github.com/openjournals/joss/issues/565).
