### 2019-08-18 #1659 @danielskatz
Also, while we wait for that, you might want to make some fixes in the bib part of the paper - anything in a title or booktitle that should be in Upper Case needs to be protected with {}s, for example
booktitle={Proc. {IEEE} Intl. Conf. on Acoustics, Speech and Signal Processing (ICASSP)},
could be
booktitle={{Proc. IEEE Intl. Conf. on Acoustics, Speech and Signal Processing (ICASSP)}},
to preserve all cases, or you could manually select parts to preserve, noting that the "." triggers automatic capitalization of the following word...

Another example is
  title        = {MUSDB18-HQ - an uncompressed version of MUSDB18},
which could be
  title        = {{MUSDB18-HQ} - an uncompressed version of {MUSDB18}},
or
  title        = {{MUSDB18-HQ - an uncompressed version of MUSDB18}},
