### 2019-08-18 #1659 @faroit
> Also, while we wait for that, you might want to make some fixes in the bib part of the paper

sure, will do.

Another question: can you please point me JOSSs pre-print policy? E.g. can I simultaneously submit to arxiv?
