### 2019-08-18 #1659 @danielskatz
This is fine, although it's a little unusual - JOSS papers are usually short and focus on the software itself, while authors often write longer papers about the use of the software, for journals or conferences, and those longer papers are the ones that are usually submitted as pre-prints.
