### 2019-08-19 #1659 @faroit
> This is fine, although it's a little unusual - JOSS papers are usually short and focus on the software itself, while authors often write longer papers about the use of the software, for journals or conferences, and those longer papers are the ones that are usually submitted as pre-prints.

maybe our contribution is a bit unusual here, indeed. We actually wrote the paper to be significanlty  longer, describing the technical details of the model (see [arxiv branch](https://github.com/sigsep/open-unmix-paper-joss/tree/arxiv)) and then shortened it to <2000 words to comply with the JOSS policy.
Since there are not many scientific contributions being made – the model is basically the distillation of multiple former methods, we therefore, really would like to stick to a JOSS-only contribution for now and will probably write a longer journal later this year, that include larger scale experiments, hence a more rigor scientific contribution.

Now, we still want to offer a DOI for the community as quickly as possible, because we expect them to cite this paper very soon with the upcoming deadlines of the major conferences in our field.
