### 2019-08-19 #1659 @faroit
Researchers close to field (audio+ML) with open-source software experience: @bmcfee  @justinsalamon @hagenw @danstowell @drscotthawley @keunwoochoi @jordipons
