### 2019-08-19 #1659 @arokem
👋 @justinsalamon : would you be willing to review this paper for the [Journal of Open Source Software](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html)?
