### 2019-08-21 #1659 @arokem
To facilitate a quick start for the review, let's try the shotgun approach:

@hagenw @danstowell @drscotthawley @keunwoochoi @jordipons (and @justinsalamon, in case you are interested):

Would one of you be willing to review this paper for the [Journal of Open Source Software](https://joss.theoj.org/)?
