### 2019-08-21 #1659 @hagenw
I would be willing
