### 2019-08-21 #1667 @whedon
[REVIEW]: Open-Unmix - A Reference Implementation for Music Source Separation
**Submitting author:** @faroit (<a href="http://orcid.org/0000-0002-2534-1165">Fabian-Robert Stöter</a>)
**Repository:** <a href="https://github.com/sigsep/open-unmix-paper-joss" target ="_blank">https://github.com/sigsep/open-unmix-paper-joss</a>
**Version:** v1.0.0
