### 2019-08-21 #1659 @arokem
OK, @bmcfee and @hagenw : please head over to #1667 for your review checklists. I'll close this issue now.
