### 2019-08-21 #1667 @bmcfee
@arokem just checking over the COI policy -- the main author has contributed to some software projects that i maintain (and have published on).  It's not totally obvious to me whether that constitutes a COI.  I don't think this would impede my ability to review, but full disclosure and all.
