### 2019-08-21 #1667 @arokem
Just to clarify: has the main author been a co-author on any of the papers that you have published on your software?
