### 2019-08-22 #1667 @arokem
In my reading, this does not constitute COI. If you feel at all uncomfortable, I'd be happy to look for another reviewer. Otherwise, I think that your statement that this association will not impede your ability to review the article is sufficient for you to check that box and move ahead.
