### 2019-08-22 sigsep/open-unmix-paper-joss#2 @bmcfee
JOSS review: version numbers and github releases
(Note: I'll be opening a bunch of issues relating to JOSS review.)

The review form asks if version number for the software reflects that reported in the github release.  There do not appear to be any release tags for this project, and the repository (repositories?) do not contain version identifiers.
