### 2019-08-22 #1667 @bmcfee
https://github.com/sigsep/open-unmix-paper-joss/issues/2 - on version numbers
