### 2019-08-22 sigsep/open-unmix-paper-joss#3 @bmcfee
JOSS review: source code links?
The JOSS review template asks if the source code is included in the submitted repo (ie this one, for the paper).

I think the answer here is broadly yes, by way of submodules, but it's not obvious at a glance which of the submodules are functional and which are stubs for future work.

If the README for this repo had an index that clarified this, it would be helpful.

As an aside, the readme in the pytorch implementation (and likely others) links to related projects denoted as "torch", "tensorflow", "nnabla", etc.  This is misleading, since they link to your repositories containing those respective implementations of UMX, and not the torch, tf, or nn projects themselves.  I suggest renaming those links.
