### 2019-08-22 #1667 @bmcfee
https://github.com/sigsep/open-unmix-paper-joss/issues/3 - on including source code and a related issue with the docs
