### 2019-08-22 sigsep/open-unmix-paper-joss#4 @bmcfee
JOSS review: evaluation
JOSS review form asks if reported performance metrics are substantiated.  I tried to run the evaluation command listed in the (torch) repo, and got the following error:

```
  File "eval.py", line 159, in <module>
    results.add_track(scores)
  File "/home/bmcfee/miniconda3/envs/open-unmix-pytorch-linux-cpu/lib/python3.7/site-packages/museval/aggregate.py", line 184, in add_track
    self.df = self.df.append(track.df, ignore_index=True)
  File "/home/bmcfee/miniconda3/envs/open-unmix-pytorch-linux-cpu/lib/python3.7/site-packages/museval/aggregate.py", line 114, in df
    return json2df(simplejson.loads(self.json), self.track_name)
  File "/home/bmcfee/miniconda3/envs/open-unmix-pytorch-linux-cpu/lib/python3.7/site-packages/museval/aggregate.py", line 414, in json2df
    df.drop(['metrics'], axis=1),
  File "/home/bmcfee/miniconda3/envs/open-unmix-pytorch-linux-cpu/lib/python3.7/site-packages/pandas/core/frame.py", line 4097, in drop
    errors=errors,
  File "/home/bmcfee/miniconda3/envs/open-unmix-pytorch-linux-cpu/lib/python3.7/site-packages/pandas/core/generic.py", line 3915, in drop
    obj = obj._drop_axis(labels, axis, level=level, errors=errors)
  File "/home/bmcfee/miniconda3/envs/open-unmix-pytorch-linux-cpu/lib/python3.7/site-packages/pandas/core/generic.py", line 3947, in _drop_axis
    new_axis = axis.drop(labels, errors=errors)
  File "/home/bmcfee/miniconda3/envs/open-unmix-pytorch-linux-cpu/lib/python3.7/site-packages/pandas/core/indexes/base.py", line 5333, in drop
    raise KeyError("{} not found in axis".format(labels[mask]))
KeyError: "['metrics'] not found in axis"
```
