### 2019-08-22 sigsep/open-unmix-paper-joss#5 @bmcfee
JOSS review: documentation
The JOSS review asks about functionality documentation.  The readme is great on this, at least from the command-line perspective.  However, I couldn't find any documentation explaining how to use programmatically.  Is there no external API for this software?
