### 2019-08-22 sigsep/open-unmix-paper-joss#6 @bmcfee
JOSS review: community guidelines
The JOSS reviewer instructions list the following points for community guidelines:

-    Contribute to the software
-    Report issues or problems with the software
-    Seek support

There is a bit in the readme addressing the first point, but the second and third points are not addressed.  In general, a contributing.md document and issue templates would be great additions.
