### 2019-08-22 #1667 @bmcfee
https://github.com/sigsep/open-unmix-paper-joss/issues/6 - on community guidelines
