### 2019-08-24 sigsep/open-unmix-paper-joss#4 @faroit
The bug occured in `museval`, caused by a pandas dependency update of `v0.25`. The bug was quickly [fixed by a community member](https://github.com/sigsep/sigsep-mus-eval/pull/64).

Since then, I released a stable version of museval and froze the requirement for open-unmix. Things should be fine now.
