### 2019-08-24 sigsep/open-unmix-paper-joss#5 @faroit
Thanks for mentioning this point. In fact we don't know what the best way would be to implement this.

One the one hand, we explicitly designed this as a non-package, being close to a "MNIST-style" example such as being done in [pytorch's example repo](https://github.com/pytorch/examples/tree/master/mnist). Furthermore, I don't think it is common to offer an API for the training or dataset creation.

On the other hand users could of course interface the `test.py` (which, I guess, would be the most common thing users would want to do) but due to its simplicity of the code, instead of having a separate documentation, I would propose to:

* add a clear `inference.md/test.md` to the [`docs` folder](https://github.com/sigsep/open-unmix-pytorch/tree/master/docs)
* add docstring to the separate function
* extend the readme a bit to point to these files

We definitely plan to make a inference python package similar to [crepe](https://github.com/marl/crepe), but we want to postpone this when we released the tf version to see which version runs faster and is easier to deploy.

Would these changes be sufficient?
