### 2019-08-25 sigsep/open-unmix-paper-joss#5 @bmcfee
> Would these changes be sufficient?

Yeah, I think so.  It's reasonable enough to make the command-line API the primary interface, but it's a little unusual in python land.  As long as it's clearly documented that this is how the software is intended to be used, I think that's fine.
