### 2019-08-27 sigsep/open-unmix-paper-joss#3 @faroit
thanks for this comment

> I think the answer here is broadly yes, by way of submodules, but it's not obvious at a glance which of the submodules are functional and which are stubs for future work.

I added a description to the readme which submodules are finished and which are under development. I think this is related to #2

> As  an aside, the readme in the pytorch implementation (and likely others)  links to related projects denoted as "torch", "tensorflow", "nnabla",  etc.  This is misleading, since they link to your repositories  containing those respective implementations of UMX, and not the torch,  tf, or nn projects themselves.  I suggest renaming those links.

I have updated the breadcrumb in all of the related repos
