### 2019-08-27 sigsep/open-unmix-paper-joss#5 @faroit
>  It's reasonable enough to make the command-line API the primary interface, but it's a little unusual in python land.

I think non-packaged code are actually quite common for simple reference implementations. I mainly followed popular repositories such as [pytorch-CycleGAN-and-pix2pix](https://github.com/junyanz/pytorch-CycleGAN-and-pix2pix) which seems to hit the right spot between simple to extend and being reproducible.

> As long as it's clearly documented that this is how the software is intended to be used, I think that's fine.

* We updated `README.md` and added `/docs/inference.md` in the [latest pytorch version](https://github.com/sigsep/open-unmix-pytorch/tree/469c0742cbafb285108543a432bfaa4b6a3f0dac)

* We updated the docstrings for the separation function. See [here](https://github.com/sigsep/open-unmix-pytorch/blob/b48f377f4a42f812a2437929b498b84fbb76a827/test.py#L84-L132)
