### 2019-08-27 sigsep/open-unmix-paper-joss#6 @faroit
thanks for the comment.

The mentioned points have been addressed in the [pytorch submodule](https://github.com/sigsep/open-unmix-pytorch/tree/469c0742cbafb285108543a432bfaa4b6a3f0dac). The contribution guidelines will be added to the nnabla and tensorflow modules once they are stable.

Regarding `seek support`, we have not explicitly addressed this point yet. Can you give an example of how we can phrase that __support is done via github issues__?
