### 2019-08-27 sigsep/open-unmix-paper-joss#2 @faroit
I think this a bit tricky to address and we would love to have your comments on this. As you pointed out there are currently no version identifiers whatsoever. This is mainly because I didn't want to confuse users with too many different versions all over the place. Currently we are facing the following situation:

The pre-trained models are hosted and versioned on zenodo. E.g. the `umx` models is deposited [here](https://zenodo.org/deposit/3370486) and I think it is the best to use releases for code that matches pre-trained models. In the model files we state which commit was used to train the model, is this particular case it was [43372f7](https://github.com/sigsep/open-unmix-pytorch/tree/43372f7d58a5d150b89ee6bf378297713def59ed). This is an old version of the repo which had significant shortcomings with respect to the documentation or data loaders other than musdb18, so we don't want to tag this `v1.0.0`. However, the most recent version is still compatible with respect to the model and the weights. Now, if we might update the model in the future - thus introducing an incompatible change, we will retrain the model and tag a new release.
Furthermore, torchhub suggests to use tags when performing `torch.hub.load('tag')`, so this should match to the other tags as well.

Therefore it would be the best to tag the pytorch version with `v1.0.0` as soon as this JOSS review is nearly completed, leaving the nnabla and tensorflow versions (which are not planned for a pre-trained model) to no-tag or tagged  as `0.1-dev`.

Thanks for your input on this.
