### 2019-08-27 sigsep/open-unmix-paper-joss#2 @bmcfee
Yeah, that's indeed trickier than the typical OSS package release.  I do think it's important to have versioning for citation and replication purposes, and your proposed scheme seems good to me.  It's maybe worth looping in @arokem for a second opinion on how this aligns to the spirit of the version requirements in JOSS submissions.
