### 2019-08-27 sigsep/open-unmix-paper-joss#6 @bmcfee
> Can you give an example of how we can phrase that **support is done via github issues**?

I can't think of any off-hand, but I think a link to the issues page with a comment to the effect of "bug reports and requests for technical support should be submitted through github issues" would be sufficient.
