### 2019-08-28 sigsep/open-unmix-paper-joss#2 @arokem
For the purpose of the paper, I think that clarity about which version of each one of the sub-modules is referenced would be important and archival versions for each one of them would need to be created. I think this means creating a release/tag and zenodo archive for each one of them when the paper is close to acceptance and then pointing from your README here to these DOIs. Then, once the paper is accepted, you create one more archive for this repo, including that README, with the DOIs pointing to the archival versions of each sub-module. Does that make sense?
