### 2019-08-28 sigsep/open-unmix-paper-joss#2 @faroit
@arokem yes, I think I understand.

In detail, JOSS is okay with the fact that one particular submodule (namely the tensorflow reimplementation) will be tagged and archived without being fully functional (e.g. a `0.1-dev` release) and the release tag of the JOSS submission will match to our main/lead implementation. We will take care that this is explictly mentioned in the readme of this repo as well as the actual paper. Is this suffcient?
