### 2019-08-29 sigsep/open-unmix-paper-joss#2 @arokem
Yes. I think this is fine as long as it's clear (both in the docs here, as well as in the paper) what works and what doesn't.
