### 2019-08-29 sigsep/open-unmix-paper-joss#2 @hagenw
I'm also a little bit puzzled by combining different implementations/sub-modules into one submission for a software paper. I see the reasoning, as the goal of all sub-modules is to implement the same functionality.

What I cannot see at the moment, is why it is sufficient for the paper when only one of those implementations is ready (`torch`). In my opinion it would make more sense to focus the paper on the `torch` implementation, or submit it when all implementations are ready.

To cite from https://joss.theoj.org/about#submitting:
> JOSS submissions must:
> [...]
> * Be feature-complete (no half-baked solutions) and be designed for maintainable extension (not one-off modifications).

@faroit: do you also plan to maintain all three different implementations, or will you not most likely benchmark them and in the end continue with the best one?
