### 2019-08-29 sigsep/open-unmix-paper-joss#9 @hagenw
Tests for the different implementations vary
The implementation of [open-unmix-pytorch](https://github.com/sigsep/open-unmix-pytorch) contains a reasonable amount of nice tests.

The same cannot be stated for [open-unmix-nnabla](https://github.com/sigsep/open-unmix-nnabla) which only contains a test for `stft`, and [open-unmix-tensorflow](https://github.com/sigsep/open-unmix-tensorflow), which does not contain any code/tests (see also https://github.com/sigsep/open-unmix-paper-joss/issues/2#issuecomment-526094678).

In my opinion this is fine if the structure of the paper would change with a focus on [open-unmix-pytorch](https://github.com/sigsep/open-unmix-pytorch) and mentioning the other packages only as additions. Otherwise I would expect a similar amount of tests in all three packages.
