### 2019-08-29 sigsep/open-unmix-paper-joss#2 @faroit
@hagenw thanks for your input.

I agree that a feature-complete submission for all three implementations would be optimal. But at least for tensorflow we don't see the benefit to release a 1.x version when so many things change with 2.0 (which also concern audio data loading as this is important for us).

> @faroit: do you also plan to maintain all three different implementations, or will you not most likely benchmark them and in the end continue with the best one?

We see the current submission addressing the scientific community first for which we think the  pytorch version is best suited. Therefore we will only release the pre-trained models for pytorch.
However, we plan to maintain all three implementations as far as framework updates concern. Also we might choose to use a tensorflow version of the model later for deployment on a interactive website. The core model will be always kept in sync though.

In the end we want researchers to be able to cite only one DOI when they use any version of open-unmix. For this to happen I see two ways:

A) We might be better off with an arxiv submission focussing on the technical details of the model, withdrawing this submission. Then we could submit separate JOSS submissions for each implementation. This, however, means that researchers would probably mainly only cite the arxiv version, which we don't think is optimal.

B) We proceed with the submission with the following changes:

* remove the tf stub submodule and remove it from the main section of the paper (but keep in the outlook section)
* wait for the nnabla version to be tested (will be this week) but not feature complete compared to the pytorch version
* tag and release both, the pytorch and nnabla version with the nnabla version not getting a `1.0` tag.

C) same as B) but also remove nnabla

@arokem @hagenw @bmcfee please tell me what you think.
