### 2019-08-29 sigsep/open-unmix-paper-joss#9 @faroit
> The same cannot be stated for open-unmix-nnabla which only contains a test for stft

For now, we cannot test the nnabla model on travis as the framework requires cudnn to run the LSTM layer.

> In my opinion this is fine if the structure of the paper would change with a focus on open-unmix-pytorch and mentioning the other packages only as additions.

See #2
