+++ paper.bib
@@ -139,6 +139,7 @@ month={May},}
   number={4},
   pages={1118--1133},
   year={2011},
+  doi={10.1109/TASL.2011.2172425},
   publisher={IEEE}
 }
 
@@ -168,6 +169,7 @@ month={May},}
   title={Monoaural Audio Source Separation Using Deep Convolutional Neural Networks},
   booktitle={Latent Variable Analysis and Signal Separation},
   year={2017},
+  doi={10.1007/978-3-319-53547-0_25},
   pages="258--266",
 }
 
@@ -180,12 +182,14 @@ month={May},}
 	Pages = {1928-1936},
 	Title = {The Signal Separation Evaluation Campaign (2007-2010): Achievements and Remaining Challenges},
 	Volume = {92},
+        doi={10.1016/j.sigpro.2011.10.007},
 	Year = {2012}}
 
 @inproceedings{sisec11,
 	Author = {Shoko Araki and Francesco Nesta and Emmanuel Vincent and Zbynek Koldovsky and Guido Nolte and Andreas Ziehe and Alexis Benichoux},
 	Booktitle = {10th International Conference on Latent Variable Analysis and Signal Separation},
 	Title = {The 2011 Signal Separation Evaluation Campaign ({SiSEC2011}): - Audio source separation -},
+        doi={10.1007/978-3-642-28551-6_51},
 	Year = {2012}}
 
 @inproceedings{sisec13,
@@ -193,6 +197,7 @@ month={May},}
 	Booktitle = {Proc. {IEEE} International Workshop on Machine Learning for Signal Processing ({MLSP})},
 	Pages = {1--6},
 	Title = {The 2013 Signal Separation Evaluation Campaign},
+        doi={10.1109/MLSP.2013.6661988},
 	Year = {2013}}
 
 @inproceedings{sisec15,
@@ -201,6 +206,7 @@ month={May},}
 	Booktitle = {Proc. Intl. Conference on Latent Variable Analysis and Signal Separation (LVA/ICA)},
 	Month = aug,
 	Title = {The 2015 Signal Separation Evaluation Campaign},
+        doi={10.1007/978-3-319-22482-4_45},
 	Year = {2015}}
 
 @inproceedings{sisec16,
@@ -209,6 +215,7 @@ month={May},}
 	Pages = {323--332},
 	Publisher = {Springer International Publishing},
 	Title = {The 2016 Signal Separation Evaluation Campaign},
+        doi={10.1007/978-3-319-53547-0_31},
 	Year = {2017}}
 
 @InProceedings{sisec18,
@@ -217,12 +224,13 @@ month={May},}
   booktitle="Latent Variable Analysis and Signal Separation:
   14th International Conference, LVA/ICA 2018, Surrey, UK",
   year="2018",
+  doi={10.1007/978-3-319-93764-9_28},
   pages="293--305"
 }
 
 @inproceedings{amodei16,
 	Author = {D. Amodei and R. Anubhai and E. Battenberg and C. Case and J. Casper and B. Catanzaro and J. Chen and M. Chrzanowski and A. Coates and G. Diamos and others},
-	Booktitle = icml,
+	Booktitle = {ICML},
 	Pages = {173--182},
 	Title = {Deep Speech 2: End-to-end speech recognition in {English and Mandarin}},
 	Year = {2016}}
@@ -245,6 +253,7 @@ month={May},}
   volume={36},
   number={1},
   pages={31-40},
+  doi={10.1109/MSP.2018.2874719},
   month=jan
 }
 
@@ -285,6 +294,7 @@ month={May},}
 	Booktitle = icassp,
 	Month = mar,
[…]
