+++ paper.bib
@@ -237,7 +237,7 @@ month={May},}
 
 @misc{rafii17,
 	Author = {Rafii, Zafar and Liutkus, Antoine and Fabian-Robert St{\"o}ter and Mimilakis, Stylianos Ioannis and Bittner, Rachel},
-	Doin = {10.5281/zenodo.1117372},
+	doi = {10.5281/zenodo.1117372},
 	Month = dec,
 	Title = {MUSDB18, a corpus for audio source separation},
 	Url = {https://sigsep.github.io/musdb},
