### 2019-08-30 #1667 @hagenw
https://github.com/sigsep/open-unmix-paper-joss/issues/9 - on testing
