### 2019-08-30 sigsep/open-unmix-paper-joss#2 @hagenw
I guess the issue comes from intermixing software and journal publications as it is of course the goal of JOSS.
* From a journal perspective the software should be feature complete, otherwise the publication could seem strange, if it is just a declaration of intend to write some software.
* From a researcher perspective that develops open source software and hopes that the community picks up, an early release of the paper is crucial as she/he can then get credits for the work in the form of citations.

I don't think that this can easily be solved besides getting rid of the paper/citations madness, but I will definitely not start that discussion here ;)

For me C) would be a good solution, but I would be also fine with the proposal of B) if this is favored by the other reviewers.
