### 2019-08-30 sigsep/open-unmix-paper-joss#2 @bmcfee
Options B or C sound fine to me.

It seems like the pytorch implementation is the first-class artifact, and would be enough on its own to merit a joss publication IMO.  Ports / implementations in other frameworks seem to me like extensions or iterations, which are obviously important to users, but not directly relevant to the pub here.  The story might be different if the focus here was on cross-framework implementation (eg onnx) rather than a specific model and application, but that's clearly not the focus of this work.
