### 2019-08-30 ab7e631
update DOIs for umx
