+++ paper.bib
@@ -67,8 +67,8 @@ month={May},}
   title        = {Open-Unmix-Pytorch UMX},
   month        = aug,
   year         = 2019,
-  doi          = {10.5281/zenodo.3340804},
-  url          = {https://doi.org/10.5281/zenodo.3340804}
+  doi          = {10.5281/zenodo.3370486},
+  url          = {https://doi.org/10.5281/zenodo.3370486}
 }
 
 @misc{umxhq,
@@ -77,8 +77,8 @@ month={May},}
   title        = {Open-Unmix-Pytorch UMX-HQ},
   month        = aug,
   year         = 2019,
-  doi          = {10.5281/zenodo.3267291},
-  url          = {https://doi.org/10.5281/zenodo.3267291}
+  doi          = {10.5281/zenodo.3370489},
+  url          = {https://doi.org/10.5281/zenodo.3370489}
 }
 
 @misc{museval,
