### 2019-08-30 3cba898
add pytorch release tag and rewrite the framework implemenations
