+++ paper.bib
@@ -61,6 +61,17 @@ month={May},}
   url          = {https://doi.org/10.5281/zenodo.3338373}
 }
 
+@misc{openunmixpytorch,
+  author       = {Fabian-Robert Stöter and
+                  Antoine Liutkus},
+  title        = {{sigsep/open-unmix-pytorch: Initial release of 
+                   Open-Unmix}},
+  month        = aug,
+  year         = 2019,
+  doi          = {10.5281/zenodo.3382104},
+  url          = {https://doi.org/10.5281/zenodo.3382104}
+}
+
 @misc{umx,
   author       = {Stöter, Fabian-Robert and
                   Antoine Liutkus},
