### 2019-08-30 sigsep/open-unmix-paper-joss#2 @faroit
@hagenw @bmcfee thanks!

We went with C)

* I updated the README.md of this project
* the paper is updated to refer to the non-pytorch version as "ports" without pre-trained models
* the pytorch version is tagged and released as `v1.0.0`

I would propose to still leave the utilities as submodules here, since they are a mandatory requirement for running open-unmix and we don't plan to submit them to JOSS at this point. I am open to remove them as well, if you think this makes more sense.
