### 2019-08-30 sigsep/open-unmix-paper-joss#9 @faroit
@hagenw  SInce a decision was made in #2, this can be closed, right?
