### 2019-08-30 sigsep/open-unmix-paper-joss#5 @faroit
@bmcfee is this sufficient?
