### 2019-08-30 sigsep/open-unmix-paper-joss#2 @bmcfee
Thanks @faroit (and all other commenters).  I think this resolves the issue, but I'll hold off on closing it out pending commentary from @hagenw and/or @arokem
