### 2019-08-31 #1667 @faroit
we would like to officially release the software next week. It would be great if we could provide a DOI soon so that researchers can use this as a baseline e.g. for the upcoming [ICASSP](https://2020.ieeeicassp.org/authors/important-dates/).

@hagenw @bmcfee thank you again for your time and help. It would be great to know for us if could tell us when you find time to complete the review so we know if we might want to prepare a preprint.
