### 2019-09-01 sigsep/open-unmix-paper-joss#9 @hagenw
Yes, this is solved by #2.
