### 2019-09-01 sigsep/open-unmix-paper-joss#11 @hagenw
JOSS review: paper proof read
The paper looks good, I have only some minor comments, that would be nice to address:

* There is a broken link for the PyTorch implementation, see https://github.com/sigsep/open-unmix-paper-joss/commit/3cba8981bafb00980ad658b4e68da35eb469cdcf#r34909982
* Some references are written with brackets when there shouldn't be one, e.g.
> ... overview see (Rafii, Liutkus, Stöter, Mimilakis, & Bittner, 2017) and (Cano,
FitzGerald, Liutkus, Plumbley, & Stöter, 2019).
> ... was presented in (Ozerov, Vincent, & Bimbot,
2011; Salaün et al., 2014).

And at several other places, would be nice to fix those.
