### 2019-09-01 sigsep/open-unmix-paper-joss#12 @hagenw
JOSS review: add community guidelines for sub-projects
I totally agree with the porposal in https://github.com/sigsep/open-unmix-paper-joss/issues/2#issuecomment-526620369 to include the other sub-projects as they are an important part of open-unmix.

In order to fulfill the community guidelines requirements asked for by JOSS, it would be nice if you could add the solution of #6 also to all the sub-projects as they are missing a `CONTRIBUTING.md`, but they would clearly benefit of having one as well.
