### 2019-09-01 #1667 @hagenw
I think that should be possible. I'm done with the review of all the written parts and will install and test the software on Monday or Tuesday. If everything works fine then the review should be finished in time.
