### 2019-09-01 sigsep/open-unmix-paper-joss#2 @hagenw
@bmcfee the solution is totally fine with me.
