### 2019-09-01 #1667 @hagenw
https://github.com/sigsep/open-unmix-paper-joss/issues/11
