### 2019-09-01 sigsep/open-unmix-paper-joss#2 @arokem
Yep. Looks good!
