### 2019-09-02 sigsep/open-unmix-paper-joss#12 @faroit
Hi, we added this to `norbert`, `museval` and `musdb`. I decided to not do a issue templates for them since we don't except much issues there, so we can probably manage without.

I didn't update the tags since I believe it makes sense that they match to the zenodo release.
