### 2019-09-02 sigsep/open-unmix-paper-joss#11 @faroit
I addressed the first bullet, about the second one: I think this is the way how the JOSS templates handles multiple reference citations and I think `[@author1; @author2]` is the correct way to cite here in markdown. But I am happy to change this if there is a different option.
