### 2019-09-02 sigsep/open-unmix-paper-joss#11 @hagenw
I was not concerned about how the two authors are cited, but when you address a citation direct in text there shouldn't be any brackets around the whole block, but only around the year:
```
... overview see (Rafii, Liutkus, Stöter, Mimilakis, & Bittner, 2017) and (Cano,
FitzGerald, Liutkus, Plumbley, & Stöter, 2019)
```
should be
```
... overview see Rafii, et al. (2017) and Cano, et al. (2019)
```
but I guess you are right, and this is not possible with markdown.
