### 2019-09-02 sigsep/open-unmix-paper-joss#12 @hagenw
Cool, thank you, this is totally fine with me.
