### 2019-09-02 sigsep/open-unmix-paper-joss#11 @hagenw
For me it's fine to stay with the current paper.
