### 2019-09-03 b9c430b
Wording suggestions
A few changes to the wording of the paper.
