+++ paper.md
@@ -49,22 +49,22 @@ To address the problem, researchers exploited additional knowledge about the way
 Many of these methods were hand-crafted and tuned to a small number of music recordings [@sisec10; @sisec11; @sisec13].
 Systematic objective evaluation of these methods, however, was hardly feasible as freely available datasets did not exist at that time.
 In fact, for a meaningful evaluation, the ground truth separated stems are necessary.
-However, these are considered precious assets in the music recording industry, therefore, are usually not available because commercial music is well known to be subject to copyright protection laws.
+However, because commercial music is usually subject to copyright protection, and the separated stems are considered to be valuable assets in the music recording industry, they are usually unavailable.
 
-Nonetheless, in the past five years, freely available datasets were released that enabled the development of data-driven methods thanks to some artists who choose compatible licenses like Creative Commons.
-Since then, progress in performance was closely linked to the availability of more data that allowed the use of machine-learning-based methods.
+Nonetheless, thanks to some artists who choose licenses like Creative Commons, that allow sharing of the stems, freely available datasets were released in the past five years and have enabled the development of data-driven methods.
+Since then, progress in performance has been closely linked to the availability of more data that allowed the use of machine-learning-based methods.
 This led to a large performance boost similar to other audio tasks such as automatic speech recognition (ASR) where a large amount of data was available.
 In fact, in 2016 the speech recognition community had access to datasets with more than 10000 hours of speech [@amodei16].
 In contrast, at the same time, the _MUSDB18_ dataset was released [@rafii17] which comprises 150 full-length music tracks – a total of just 10 hours of music.
 To date, this is still the largest freely available dataset for source separation.
 Nonetheless, even with this small amount of data, deep neural networks (DNNs) were not only successfully used for music separation but they are now setting the state-of-the-art in this domain as can be seen by the results of the community-based signal separation evaluation campaign (SiSEC) [@sisec15; @sisec16; @sisec18]. In these challenges, the proposed systems are compared to other methods. Among the systems under test, classical signal processing based methods were clearly outperformed by machine learning methods. However they were still useful as a _fast_ and often _simple to understand_ baseline.
 
-In the following we will enlist a number of these reference implementations for source separation.
-While today some commercial systems also exist, such as _Audionamix XTRAX STEMS_, _IZOTOPE RX 7_ or _AudioSourceRE_ which target end-users, we only considered tools that are available as open-source software, suitable for research.
+In the following, we will describe a number of these reference implementations for source separation.
+While there are some commercial systems available, such as _Audionamix XTRAX STEMS_, _IZOTOPE RX 7_ or _AudioSourceRE_, we only considered tools that are available as open-source software, and are suitable for research.
 
 <!-- list of methods -->
 The first publicly available software for source separation was _openBlissart_, released in 2011 [@weninger11]. It is written in C++ and accounts for the class of systems that are based on non-negative matrix factorization (NMF). In 2012, the _Flexible Audio Source Separation Toolbox (FASST)_ was presented in [@ozerov2011general; @salaun14]. It is written in MATLAB/C++ and is also based on NMF methods, but also includes other model-based methods.
-In 2016, the _untwist_ library was proposed in [@roma16]. It comprises several methods, ranging from classical signal-processing-based methods to feed-forward neural networks. The library is written in Python 2.7. Unfortunately, it has not been updated since two years ago and many of its methods are not tested.
[…]
