+++ paper.bib
@@ -28,11 +28,11 @@ month={May},}
 @misc{norbert,
   author       = {Fabian-Robert {Stöter} and
                   Antoine Liutkus},
-  title        = {{sigsep/norbert: v0.2.0}},
-  month        = jul,
+  title        = {{sigsep/norbert: v0.2.1}},
+  month        = sep,
   year         = 2019,
-  doi          = {10.5281/zenodo.3269749},
-  url          = {https://doi.org/10.5281/zenodo.3269749}
+  doi          = {10.5281/zenodo.3386463},
+  url          = {https://doi.org/10.5281/zenodo.3386463}
 }
 
 @ARTICLE{McFee19,
