+++ paper.bib
@@ -26,8 +26,8 @@ month={May},}
 	Year = {1997}}
 
 @misc{norbert,
-  author       = {Fabian-Robert {Stöter} and
-                  Antoine Liutkus},
+  author       = {Antoine Liutkus and
+                  Fabian-Robert {Stöter}},
   title        = {{sigsep/norbert: v0.2.1}},
   month        = sep,
   year         = 2019,
