### 2019-09-06 #1667 @faroit
@arokem I did some last little update on the bibliography and I think we are good here. Please tell me what the next steps are to proceed.
