### 2019-09-07 #1667 @arokem
@faroit : At this point, could you please make a new release of the main repo that includes all of the changes that have resulted from the review? Then, please make an archive of the software in Zenodo/figshare/other service and update this thread with the DOI of the archive. For the archive version, please make sure that:

- The authors of the archive are the same as the JOSS paper authors
