### 2019-09-08 #1667 @arokem
@openjournals/joss-eics : I believe this is ready for your review
